Split the `zizq top` UI code into smaller modules.
All the UI rendering code for `zizq top` was in a single `ui.rs` module
which we have already outgrown. Splitting the module into smaller
modules grouped inside a `ui/` module directory makes it much easier to
navigate the code and read diffs.
